Offline-SST-System/requirements.txt
pyaudio==0.2.14
openai-whisper
torch
numpy